SOLID : astro bird
