+++ test/examples/flappy_bird/simulation-shared/observation/observation.ts
@@ -1,0 +1,9 @@
+/**
+ * Shared observation public entry.
+ *
+ * This focused boundary keeps observation feature assembly separate from
+ * observation-vector projection so browser, environment, and worker callers can
+ * depend on a smaller, clearer public surface.
+ */
+export * from './observation.features.utils';
+export * from './observation.vector.utils';
